Remove k6 browser's browser.on method doc (#1197)
Due to the changes made to k6 browser's API, for which the browser's
launch and close methods are no longer delegated/exposed to end users,
but instead handled automatically by the extension's implementation
itself, the browser.on() method is no longer applicable, as it was only
supporting the 'disconnect' event type. Therefor it is no longer exposed
to the JS API and removed from the documentation.